project/requirements.txt
langchain>=0.2.14
langchain-community>=0.2.12
langgraph>=0.2.3
langchain-google-genai>=1.0.8
faiss-cpu>=1.8.0.post1
python-dotenv>=1.0.1
streamlit>=1.37.0
